pyproject.toml
[project]
name = "assist-time-series-analysis-app"
version = "0.1.0"
description = ""
authors = [
    {name = "ruo-lee",email = "comsa333@gmail.com"}
]
readme = "README.md"
requires-python = "3.10.16"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "numpy (>=1.25.0,<2.0.0)",
    "pmdarima (>=2.0.4,<3.0.0)",
    "prophet (>=1.1.6,<2.0.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
